+++ activities/FractionBounce.activity/js/activity.js
@@ -49,6 +49,8 @@ let app = new Vue({
 		score: 0,
 		count: 0,
 		ballSrc: 'images/soccerball.svg',
+		backgroundSrc: 'url(images/grass_background.png)',
+		backgroundSize: 'cover',
 		canvasWidth: 0,
 		canvasHeight: 0,
 		l10n: {
@@ -149,6 +151,12 @@ let app = new Vue({
 				this.changeGameState();
 				this.isFirstPauseClick = true;
 			}
+		},
+		backgroundSrc: function (newVal) {
+			document.body.style.backgroundImage = newVal;
+		},
+		backgroundSize: function (newVal) {
+			document.body.style.backgroundSize = newVal;
 		}
 	},
 
@@ -165,7 +173,20 @@ let app = new Vue({
 
 		onJournalDataLoaded: function (data, metadata) {
 			console.log("Existing instance");
-			this.userFractions = data.userFractions;
+			this.userFractions = data.userFractions || [];
+			if (data.ballSrc) {
+				this.ballSrc = data.ballSrc;
+			}
+			if (data.backgroundSrc) {
+				this.backgroundSrc = data.backgroundSrc;
+			}
+			if (data.backgroundSize) {
+				this.backgroundSize = data.backgroundSize;
+			}
+			if (data.mode) {
+				this.mode = data.mode;
+			}
+			this.init();
 		},
 
 		onJournalLoadError: function (error) {
@@ -232,6 +253,7 @@ let app = new Vue({
 			}
 
 			// start by clicking ball
+			document.getElementById('slopeCanvas').removeEventListener('click', this.startGame);
 			document.getElementById('slopeCanvas').addEventListener('click', this.startGame);
 		},
 
@@ -606,20 +628,21 @@ let app = new Vue({
 
 			this.ballSrc = 'images/' + event.ball + '.svg';
 			this.img.src = this.ballSrc;
+			this.backgroundSize = 'cover';
 			switch (event.ball) {
 				case 'rugbyball':
 				case 'soccerball':
-					document.body.style.backgroundImage = 'url(images/grass_background.png)';
+					this.backgroundSrc = 'url(images/grass_background.png)';
 					break;
 				case 'bowlingball':
 				case 'basketball':
-					document.body.style.backgroundImage = 'url(images/parquet_background.png)';
+					this.backgroundSrc = 'url(images/parquet_background.png)';
 					break;
 				case 'feather':
-					document.body.style.backgroundImage = 'url(images/feather_background.png)';
+					this.backgroundSrc = 'url(images/feather_background.png)';
 					break;
 				case 'beachball':
-					document.body.style.backgroundImage = 'url(images/beach_background.png)';
+					this.backgroundSrc = 'url(images/beach_background.png)';
 					break;
 			}
 		},
@@ -630,19 +653,19 @@ let app = new Vue({
 				this.insertImage('bg');
 				return;
 			}
-			document.body.style.backgroundSize = 'cover';
+			this.backgroundSize = 'cover';
 			switch (event.bg) {
 				case 'grass':
-					document.body.style.backgroundImage = 'url(images/grass_background.png)';
+					this.backgroundSrc = 'url(images/grass_background.png)';
 					break;
 				case 'wood':
-					document.body.style.backgroundImage = 'url(images/parquet_background.png)';
+					this.backgroundSrc = 'url(images/parquet_background.png)';
 					break;
 				case 'clouds':
-					document.body.style.backgroundImage = 'url(images/feather_background.png)';
+					this.backgroundSrc = 'url(images/feather_background.png)';
 					break;
 				case 'sand':
-					document.body.style.backgroundImage = 'url(images/beach_background.png)';
+					this.backgroundSrc = 'url(images/beach_background.png)';
 					break;
 			}
 		},
@@ -654,6 +677,7 @@ let app = new Vue({
 					journalchooser.show(function (entry) {
 						if (!entry) {
 							if (to == 'ball') {
+								vm.ballSrc = data;
 								vm.img.onload();
 							}
 							return;
@@ -664,8 +688,8 @@ let app = new Vue({
 								vm.ballSrc = data;
 								vm.img.src = data;
 							} else if (to == 'bg') {
-								document.body.style.backgroundImage = 'url(' + data + ')';
-								document.body.style.backgroundSize = 'contain';
+								vm.backgroundSrc = 'url(' + data + ')';
+								vm.backgroundSize = 'contain';
 							}
 						});
 					}, { mimetype: 'image/png' }, { mimetype: 'image/jpeg' });
@@ -764,7 +788,11 @@ let app = new Vue({
 		onStop: function () {
 			// Save current library in Journal on Stop
 			var context = {
-				userFractions: this.userFractions
+				userFractions: this.userFractions,
+				ballSrc: this.ballSrc,
+				backgroundSrc: this.backgroundSrc,
+				backgroundSize: this.backgroundSize,
+				mode: this.mode
 			}
 			this.$refs.SugarJournal.saveData(context);
 			this.$refs.SugarDevice.vibrate(100);